# Allow specialization of the RouterResponder