--- a/Sources/Hummingbird/Router/Trie/Trie+resolve.swift
+++ b/Sources/Hummingbird/Router/Trie/Trie+resolve.swift
@@ -16,8 +16,8 @@ import NIOCore
 
 extension RouterTrie {
     /// Resolve a path to a `Value` if available
-    @usableFromInline
-    internal func resolve(_ path: String) -> (value: Value, parameters: Parameters)? {
+    @inlinable
+    public func resolve(_ path: String) -> (value: Value, parameters: Parameters)? {
         let pathComponents = path.split(separator: "/", omittingEmptySubsequences: true)
         var pathComponentsIterator = pathComponents.makeIterator()
         var parameters = Parameters()