I've ran the `PerformanceTest` codebase two times, before and after this PR. Note that the PerformanceTest only uses 4 cores for HB2 - so that we have resource budget for regular OS use and the `wrk` client benckmark suite.

Since I made a lot of symbols inlinable, I had to make them public or internal. Public with SPI does not work with inlinable. We'll need to decide on a public API surface if we want benchmarks.

TODOs:
- [ ] Fix the router benchmarks in the Benchmarks folder.

## Before

```
Running 15s test @ http://localhost:8080
  8 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.78ms    1.35ms  47.83ms   91.64%
    Req/Sec     9.80k     1.47k   26.85k    89.11%
  1173503 requests in 15.10s, 166.75MB read
Requests/sec:  77697.81
Transfer/sec:     11.04MB
```

## After

```
Running 15s test @ http://localhost:8080
  8 threads and 128 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency     1.69ms    1.37ms  50.63ms   93.09%
    Req/Sec    10.28k   694.90    13.73k    76.45%
  1233179 requests in 15.10s, 175.23MB read
Requests/sec:  81659.66
Transfer/sec:     11.60MB
```