@adam-fowler 2024-05-09
Can you get the Benchmarks and Tests working again, to indicate what we would need to make public. I guess for the tests it doesn't matter too much as you can import as `@testable`.